+++ R/options.R
@@ -8,7 +8,7 @@
 
 .onAttach <- function(libname, pkgname){
   
-  attempt <- try(googleAuthR::gar_attach_auto_auth("https://www.googleapis.com/auth/bigquery",
+  attempt <- try(googleAuthR::gar_attach_auto_auth("https://www.googleapis.com/auth/cloud-platform",
                                                    environment_var = "BQ_AUTH_FILE",
                                                    travis_environment_var = "TRAVIS_BQ_AUTH_FILE"))