+++ NEWS.md
@@ -15,6 +15,7 @@
 * Migrate to using `googleCloudStorageR` for Cloud Storage stuff
 * Set default authentication scope to `https://www.googleapis.com/auth/cloud-platform`
 * Unit tests
+* Upload table will now correctly report errors 
 
 # bigQueryR 0.2.0